Fixed comments and moved tests in pocket_ic
