+++ Cargo.toml
@@ -244,7 +244,6 @@ members = [
     "rs/protobuf",
     "rs/protobuf/generator",
     "rs/pocket_ic_server",
-    "rs/pocket_ic_server/test_canister",
     "rs/registry/admin",
     "rs/registry/admin-derive",
     "rs/registry/canister",
