+++ packages/pocket-ic/src/lib.rs
@@ -1,5 +1,4 @@
 #![allow(clippy::test_attr_in_doctest)]
-use crate::common::rest::TickConfigs;
 /// # PocketIC: A Canister Testing Platform
 ///
 /// PocketIC is the local canister smart contract testing platform for the [Internet Computer](https://internetcomputer.org/).
@@ -526,7 +525,7 @@ impl PocketIc {
     /// Make the IC produce and progress by one block with custom
     /// configs for the round.
     #[instrument(skip(self), fields(instance_id=self.pocket_ic.instance_id))]
-    pub fn tick_with_configs(&self, configs: TickConfigs) {
+    pub fn tick_with_configs(&self, configs: crate::common::rest::TickConfigs) {
         let runtime = self.runtime.clone();
         runtime.block_on(async { self.pocket_ic.tick_with_configs(configs).await })
     }
