+++ rs/pocket_ic_server/BUILD.bazel
@@ -179,7 +179,6 @@ rust_test(
     aliases = {},
     data = [
         ":pocket-ic-server",
-        "//rs/pocket_ic_server/test_canister:test_canister.wasm",
         "//rs/registry/canister:registry-canister",
         "@ii_dev_canister//file",
         "@mozilla_root_ca_store//file",
@@ -190,7 +189,6 @@ rust_test(
         "REGISTRY_WASM": "$(rootpath //rs/registry/canister:registry-canister)",
         "II_WASM": "$(rootpath @ii_dev_canister//file)",
         "SSL_CERT_FILE": "$(rootpath @mozilla_root_ca_store//file)",
-        "TEST_WASM": "$(rootpath //rs/pocket_ic_server/test_canister:test_canister.wasm)",
     },
     tags = [
         "cpu:8",
