+++ rs/pocket_ic_server/src/pocket_ic.rs
@@ -882,7 +882,7 @@ impl PocketIc {
 
         // We execute a round on every subnet to make sure it has a state to certify.
         for subnet in subnets.get_all() {
-            subnet.state_machine.execute_round(None);
+            subnet.state_machine.execute_round();
         }
 
         // We initialize delegation from NNS.
@@ -1622,21 +1622,21 @@ impl Tick {
 
 impl Operation for Tick {
     fn compute(&self, pic: &mut PocketIc) -> OpOut {
-        let blockmakers_per_subnet = self.configs.blockmakers.as_ref().map(|blockmaker_configs| {
-            blockmaker_configs
-                .blockmakers_per_subnet
+        let blockmakers_per_subnet = self.configs.blockmakers.as_ref().map(|cfg| {
+            cfg.blockmakers_per_subnet
                 .iter()
-                .map(|subnet| SubnetBlockmakerMetrics::from(subnet.clone()))
+                .cloned()
+                .map(SubnetBlockmakerMetrics::from)
                 .collect_vec()
         });
-        if let Some(blockmakers_per_subnet) = blockmakers_per_subnet.as_ref() {
-            if let Err(error) = self.validate_blockmakers_per_subnet(pic, blockmakers_per_subnet) {
+
+        if let Some(ref bm_per_subnet) = blockmakers_per_subnet {
+            if let Err(error) = self.validate_blockmakers_per_subnet(pic, bm_per_subnet) {
                 return error;
             }
         }
 
         let subnets = pic.subnets.subnets.read().unwrap();
-
         for (subnet_id, subnet) in subnets.iter() {
             let blockmaker_metrics = blockmakers_per_subnet.as_ref().and_then(|bm_per_subnet| {
                 bm_per_subnet
@@ -1648,8 +1648,14 @@ impl Operation for Tick {
                     })
             });
 
-            subnet.state_machine.execute_round(blockmaker_metrics);
+            match blockmaker_metrics {
+                Some(metrics) => subnet
+                    .state_machine
+                    .execute_round_with_blockmaker_metrics(metrics),
+                _ => subnet.state_machine.execute_round(),
+            }
         }
+
         OpOut::NoOutput
     }
 
@@ -1665,7 +1671,7 @@ impl Operation for AdvanceTimeAndTick {
     fn compute(&self, pic: &mut PocketIc) -> OpOut {
         for subnet in pic.subnets.get_all() {
             subnet.state_machine.advance_time(self.0);
-            subnet.state_machine.execute_round(None);
+            subnet.state_machine.execute_round();
         }
         OpOut::NoOutput
     }
@@ -1775,7 +1781,7 @@ impl Operation for AwaitIngressMessage {
                         _ => {}
                     }
                     for subnet_ in pic.subnets.get_all() {
-                        subnet_.state_machine.execute_round(None);
+                        subnet_.state_machine.execute_round();
                     }
                 }
                 OpOut::Error(PocketIcError::BadIngressMessage(format!(
@@ -2807,7 +2813,7 @@ fn route(
                     // We need to execute a round on the new subnet to make its state certified.
                     // To keep the PocketIC instance time in sync, we execute a round on all subnets.
                     for subnet in pic.subnets.get_all() {
-                        subnet.state_machine.execute_round(None);
+                        subnet.state_machine.execute_round();
                     }
                     // We initialize delegation from NNS.
                     if let Some(nns_subnet) = pic.nns_subnet() {
