--- rs/pocket_ic_server/test_canister/BUILD.bazel
@@ -1,13 +1,0 @@
-load("//bazel:canisters.bzl", "rust_canister")
-
-rust_canister(
-    name = "test_canister",
-    srcs = ["src/canister.rs"],
-    service_file = ":canister.did",
-    deps = [
-        # Keep sorted.
-        "//rs/types/management_canister_types",
-        "@crate_index//:candid",
-        "@crate_index//:ic-cdk",
-    ],
-)
