--- rs/pocket_ic_server/test_canister/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "pocket-ic-server-test-canister"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-description.workspace = true
-documentation.workspace = true
-
-[[bin]]
-name = "pocket-ic-server-test-canister"
-path = "src/canister.rs"
-
-[dependencies]
-candid = { workspace = true }
-ic-cdk = { workspace = true }
-ic-management-canister-types = { path = "../../../rs/types/management_canister_types" }
