--- rs/pocket_ic_server/test_canister/canister.did
@@ -1,19 +1,0 @@
-type NodeMetricsHistoryArgs = record {
-    subnet_id: principal;
-    start_at_timestamp_nanos: nat64;
-};
-
-type NodeMetrics = record {
-    node_id: principal;
-    num_blocks_proposed_total: nat64;
-    num_block_failures_total: nat64;
-};
-
-type NodeMetricsHistoryResponse = record {
-  timestamp_nanos: nat64;
-  node_metrics: vec NodeMetrics;
-};
-
-service : {
-  node_metrics_history_proxy: (NodeMetricsHistoryArgs) -> (vec NodeMetricsHistoryResponse);
-}
