+++ rs/pocket_ic_server/tests/test.rs
@@ -1,24 +1,20 @@
-use candid::{Decode, Encode, Principal};
+use candid::{Encode, Principal};
 use ic_agent::agent::CallResponse;
 use ic_cdk::api::management_canister::main::CanisterIdRecord;
 use ic_cdk::api::management_canister::provisional::ProvisionalCreateCanisterWithCyclesArgument;
 use ic_interfaces_registry::{
     RegistryDataProvider, RegistryVersionedRecord, ZERO_REGISTRY_VERSION,
 };
-use ic_management_canister_types::{
-    NodeMetricsHistoryArgs, NodeMetricsHistoryResponse, ProvisionalCreateCanisterWithCyclesArgs,
-};
+use ic_management_canister_types::ProvisionalCreateCanisterWithCyclesArgs;
 use ic_registry_proto_data_provider::ProtoRegistryDataProvider;
 use ic_registry_transport::pb::v1::{
     registry_mutation::Type, RegistryAtomicMutateRequest, RegistryMutation,
 };
-use ic_types::PrincipalId;
 use ic_utils::interfaces::ManagementCanister;
 use nix::sys::signal::Signal;
 use pocket_ic::common::rest::{
-    BlockMakerConfigs, CreateHttpGatewayResponse, HttpGatewayBackend, HttpGatewayConfig,
-    HttpGatewayDetails, HttpsConfig, InstanceConfig, RawSubnetBlockmakerMetrics, SubnetConfigSet,
-    SubnetKind, TickConfigs, Topology,
+    CreateHttpGatewayResponse, HttpGatewayBackend, HttpGatewayConfig, HttpGatewayDetails,
+    HttpsConfig, InstanceConfig, SubnetConfigSet, SubnetKind, Topology,
 };
 use pocket_ic::{update_candid, PocketIc, PocketIcBuilder};
 use rcgen::{CertificateParams, KeyPair};
@@ -1481,89 +1477,4 @@ fn test_gateway_address_in_use() {
         "Failed to bind to address 127.0.0.1:{}: Address already in use",
         port
     )));
-}
-
-fn test_canister_wasm() -> Vec<u8> {
-    let wasm_path = std::env::var_os("TEST_WASM").expect("Missing test canister wasm file");
-    std::fs::read(wasm_path).unwrap()
-}
-
-#[test]
-fn test_custom_blockmaker() {
-    const INIT_CYCLES: u128 = 2_000_000_000_000;
-    const HOURS_IN_SECONDS: u64 = 60 * 60;
-
-    let pocket_ic = PocketIcBuilder::new().with_application_subnet().build();
-    let topology = pocket_ic.topology();
-    let application_subnet = topology.get_app_subnets()[0];
-
-    // Create and install test canister.
-    let canister = pocket_ic.create_canister_on_subnet(None, None, application_subnet);
-    pocket_ic.add_cycles(canister, INIT_CYCLES);
-    pocket_ic.install_canister(canister, test_canister_wasm(), vec![], None);
-
-    let nodes = topology
-        .subnet_configs
-        .get(&application_subnet)
-        .unwrap()
-        .clone();
-
-    let blockmaker_1 = nodes.node_ids[0].clone();
-    let blockmaker_2 = nodes.node_ids[1].clone();
-
-    let subnets_blockmakers = vec![RawSubnetBlockmakerMetrics {
-        subnet: application_subnet.into(),
-        blockmaker: blockmaker_1.clone(),
-        failed_blockmakers: vec![blockmaker_2.clone()],
-    }];
-
-    let mut tick_configs = TickConfigs {
-        blockmakers: Some(BlockMakerConfigs {
-            blockmakers_per_subnet: subnets_blockmakers,
-        }),
-    };
-    let daily_blocks = 5;
-
-    // Blockmaker metrics are recorded in the management canister
-    for _ in 0..daily_blocks {
-        pocket_ic.tick_with_configs(tick_configs.clone());
-    }
-    // Advance time until next day so that management canister can record blockmaker metrics
-    pocket_ic.advance_time(std::time::Duration::from_secs(HOURS_IN_SECONDS * 24));
-    pocket_ic.tick();
-
-    let response = pocket_ic
-        .update_call(
-            canister,
-            Principal::anonymous(),
-            // Calls the node_metrics_history method on the management canister
-            "node_metrics_history_proxy",
-            Encode!(&NodeMetricsHistoryArgs {
-                subnet_id: PrincipalId::from(application_subnet),
-                start_at_timestamp_nanos: 0,
-            })
-            .unwrap(),
-        )
-        .unwrap();
-
-    let first_node_metrics = Decode!(&response, Vec<NodeMetricsHistoryResponse>)
-        .unwrap()
-        .remove(0)
-        .node_metrics;
-
-    let blockmaker_1_metrics = first_node_metrics
-        .iter()
-        .find(|x| x.node_id.0 == Principal::from(blockmaker_1.clone()))
-        .unwrap()
-        .clone();
-    let blockmaker_2_metrics = first_node_metrics
-        .into_iter()
-        .find(|x| x.node_id.0 == Principal::from(blockmaker_2.clone()))
-        .unwrap();
-
-    assert_eq!(blockmaker_1_metrics.num_blocks_proposed_total, daily_blocks);
-    assert_eq!(blockmaker_1_metrics.num_block_failures_total, 0);
-
-    assert_eq!(blockmaker_2_metrics.num_blocks_proposed_total, 0);
-    assert_eq!(blockmaker_2_metrics.num_block_failures_total, daily_blocks);
 }
