+++ rs/state_machine_tests/tests/multi_subnet.rs
@@ -145,7 +145,7 @@ fn counter_canister_call_test() {
                 .build(),
         )
         .unwrap();
-    env1.execute_round(None);
+    env1.execute_round();
     let wasm_result = env1.await_ingress(msg_id, MAX_TICKS).unwrap();
     match wasm_result {
         WasmResult::Reply(bytes) => assert_eq!(bytes, 10_000_000_u32.to_le_bytes()),
@@ -170,7 +170,7 @@ fn counter_canister_call_test() {
                 .build(),
         )
         .unwrap();
-    env1.execute_round(None);
+    env1.execute_round();
     let wasm_result = env1.await_ingress(msg_id, MAX_TICKS).unwrap();
     match wasm_result {
         // The call fails with CANISTER_ERROR reject code (5).
@@ -191,7 +191,7 @@ fn counter_canister_call_test() {
                 .build(),
         )
         .unwrap();
-    env1.execute_round(None);
+    env1.execute_round();
     let wasm_result = env1.await_ingress(msg1_id, MAX_TICKS).unwrap();
     match wasm_result {
         WasmResult::Reply(bytes) => assert_eq!(bytes, vec![42; 2000000]),
@@ -211,7 +211,7 @@ fn counter_canister_call_test() {
                 .build(),
         )
         .unwrap();
-    env2.execute_round(None);
+    env2.execute_round();
     let wasm_result = env2.await_ingress(msg2_id, MAX_TICKS).unwrap();
     match wasm_result {
         WasmResult::Reply(bytes) => assert_eq!(bytes, vec![123; 2000000]),
@@ -238,9 +238,9 @@ fn counter_canister_call_test() {
     // inter-canister call, and finally we execute a round on the 1st subnet
     // to process the callback of the inter-canister call and finish processing
     // the ingress message.
-    env1.execute_round(None);
-    env2.execute_round(None);
-    env1.execute_round(None);
+    env1.execute_round();
+    env2.execute_round();
+    env1.execute_round();
 
     let wasm_result = env1.await_ingress(msg3_id, MAX_TICKS).unwrap();
     match wasm_result {
@@ -326,7 +326,7 @@ fn counter_canister_call_test() {
 
     // This time we need to execute multiple rounds on the 1st subnet
     // to induct all ingress messages with large payloads.
-    env1.execute_round(None);
+    env1.execute_round();
     assert!(matches!(
         (
             env1.ingress_status(&msg10_id),
@@ -342,7 +342,7 @@ fn counter_canister_call_test() {
 
     // The third ingress message is only inducted after a repeated
     // call to execute a round.
-    env1.execute_round(None);
+    env1.execute_round();
     assert!(matches!(
         (
             env1.ingress_status(&msg10_id),
@@ -360,21 +360,21 @@ fn counter_canister_call_test() {
     // to induct the ingress message with large payload
     // and all three inter-canister calls with large arguments
     // from the 1st subnet.
-    env2.execute_round(None);
+    env2.execute_round();
     assert!(matches!(
         env2.ingress_status(&msg20_id),
         IngressStatus::Known { .. }
     ));
-    env2.execute_round(None);
-    env2.execute_round(None);
+    env2.execute_round();
+    env2.execute_round();
     // Finally, we need to execute multiple rounds on the 1st subnet
     // to induct all (large) responses from the 2nd subnet
     // and an inter-canister call from the 2nd into the 1st subnet
     // with large argument.
-    env1.execute_round(None);
-    env1.execute_round(None);
-    env1.execute_round(None);
-    env1.execute_round(None);
+    env1.execute_round();
+    env1.execute_round();
+    env1.execute_round();
+    env1.execute_round();
 
     let wasm_result = env1.await_ingress(msg10_id, MAX_TICKS).unwrap();
     match wasm_result {
@@ -395,7 +395,7 @@ fn counter_canister_call_test() {
     // This time, we also need to execute one more round on the 2nd subnet
     // to process the response callback of the inter-canister call
     // to the 1st subnet.
-    env2.execute_round(None);
+    env2.execute_round();
 
     let wasm_result = env2.await_ingress(msg20_id, MAX_TICKS).unwrap();
     match wasm_result {
